+++ config/proptype-checker.js
@@ -1,7 +1,7 @@
-/* global console */
+/* global console, beforeEach, afterEach, expect */
 /* eslint no-var: off, no-console: ["error", { allow: ["error"] }] */
 
-var spy = require('enyo-console-spy');
+var spy = require('console-snoop');
 
 var watchErrorAndWarnings = spy.watchErrorAndWarnings;
 var filterErrorAndWarnings = spy.filterErrorAndWarnings;
@@ -9,7 +9,7 @@ var restoreErrorAndWarnings = spy.restoreErrorAndWarnings;
 
 beforeEach(watchErrorAndWarnings);
 
-afterEach(function (done) {
+afterEach(function(done) {
 	const actual = filterErrorAndWarnings(/(Invalid prop|Failed prop type|Unknown prop)/);
 	const expected = 0;
 	restoreErrorAndWarnings();