+++ global-cli/modifiers/stats.js
@@ -1,6 +1,6 @@
 var BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
-module.exports = function(config, opts) {
+module.exports = function(config) {
 	config.plugins.push(new BundleAnalyzerPlugin({
 		analyzerMode: 'static',
 		reportFilename: 'stats.html',