+++ global-cli/modifiers/util/EnactFrameworkRefPlugin.js
@@ -34,14 +34,12 @@ function normalizePath(dir, file, compiler) {
 	}
 }
 
+// Determine if it's a NodeJS output filesystem or if it's a foreign/virtual one.
 function isNodeOutputFS(compiler) {
-	try {
-		var NodeOutputFileSystem = require('webpack/lib/node/NodeOutputFileSystem');
-		return (compiler.outputFileSystem.writeFile===NodeOutputFileSystem.prototype.writeFile);
-	} catch(e) {
-		console.error('SnapshotPlugin loader is not compatible with standalone global installs of Webpack.');
-		return false;
-	}
+	return (compiler.outputFileSystem
+			&& compiler.outputFileSystem.constructor
+			&& compiler.outputFileSystem.constructor.name
+			&& compiler.outputFileSystem.constructor.name === 'NodeOutputFileSystem');
 }
 
 function updateAppInfo(output, blob) {
@@ -66,9 +64,9 @@ function EnactFrameworkRefPlugin(opts) {
 	this.options.external = this.options.external || {};
 	this.options.external.inject = this.options.external.inject || this.options.external.path;
 
-	if(!process.env.ILIB_LOCALE_PATH) {
-		process.env.ILIB_LOCALE_PATH = path.join(this.options.external.inject, 'node_module',
-				'@enact', 'i18n', 'ilib', 'locale');
+	if(!process.env.ILIB_BASE_PATH) {
+		process.env.ILIB_BASE_PATH = path.join(this.options.external.inject, 'node_module',
+				'@enact', 'i18n', 'ilib');
 	}
 }
 module.exports = EnactFrameworkRefPlugin;
@@ -85,9 +83,9 @@ EnactFrameworkRefPlugin.prototype.apply = function(compiler) {
 	compiler.plugin('compilation', function(compilation, params) {
 		var normalModuleFactory = params.normalModuleFactory;
 		compilation.dependencyFactories.set(DelegatedSourceDependency, normalModuleFactory);
-		
-		compilation.plugin('html-webpack-plugin-alter-chunks', function(chunks, params) {
-			var chunkFiles = [ normalizePath(external.inject, 'enact.css', compiler) ];
+
+		compilation.plugin('html-webpack-plugin-alter-chunks', function(chunks) {
+			var chunkFiles = [normalizePath(external.inject, 'enact.css', compiler)];
 			if(!external.snapshot) {
 				chunkFiles.unshift(normalizePath(external.inject, 'enact.js', compiler));
 			}
@@ -96,18 +94,22 @@ EnactFrameworkRefPlugin.prototype.apply = function(compiler) {
 				names: ['enact_framework'],
 				files: chunkFiles
 			});
-
-			// Store the absolute filepath to the external framework so the PrerenderPlugin can use it
-			params.plugin.options.externalFramework = path.resolve(path.join(external.path, 'enact.js'));
 			return chunks;
 		});
+
+		if(external.snapshot && isNodeOutputFS(compiler)) {
+			compilation.plugin('webos-meta-root-appinfo', function(meta) {
+				meta.v8SnapshotFile = normalizePath(external.inject, 'snapshot_blob.bin', compiler);
+				return meta;
+			});
+		}
 	});
 
 	// Apply the Enact factory plugin to handle the require() delagation/rerouting
 	compiler.plugin('compile', function(params) {
 		params.normalModuleFactory.apply(new DelegatedEnactFactoryPlugin({
 			name: name,
-			libraries: libs,
+			libraries: libs
 		}));
 	});