+++ global-cli/modifiers/util/SnapshotPlugin.js
@@ -6,14 +6,12 @@ var
 	findCacheDir = require('find-cache-dir'),
 	chalk = require('chalk');
 
+// Determine if it's a NodeJS output filesystem or if it's a foreign/virtual one.
 function isNodeOutputFS(compiler) {
-	try {
-		var NodeOutputFileSystem = require('webpack/lib/node/NodeOutputFileSystem');
-		return (compiler.outputFileSystem.writeFile===NodeOutputFileSystem.prototype.writeFile);
-	} catch(e) {
-		console.error('SnapshotPlugin loader is not compatible with standalone global installs of Webpack.');
-		return false;
-	}
+	return (compiler.outputFileSystem
+			&& compiler.outputFileSystem.constructor
+			&& compiler.outputFileSystem.constructor.name
+			&& compiler.outputFileSystem.constructor.name === 'NodeOutputFileSystem');
 }
 
 function getBlobName(args) {
@@ -25,19 +23,6 @@ function getBlobName(args) {
 	return 'snapshot_blob.bin';
 }
 
-function updateAppInfo(output, blob) {
-	var appInfo = path.join(output, 'appinfo.json');
-	if(exists(appInfo)) {
-		try {
-			var meta = JSON.parse(fs.readFileSync(appInfo, {encoding:'utf8'}));
-			meta.v8SnapshotFile = blob;
-			fs.writeFileSync(appInfo, JSON.stringify(meta, null, '\t'), {encoding:'utf8'});
-		} catch(e) {
-			return new Error('Failed to set "v8SnapshotFile" property in appinfo.json');
-		}
-	}
-}
-
 function SnapshotPlugin(options) {
 	this.options = options || {};
 	this.options.exec = this.options.exec || process.env.V8_MKSNAPSHOT;
@@ -46,18 +31,30 @@ function SnapshotPlugin(options) {
 		'--random-seed=314159265',
 		'--startup-blob=snapshot_blob.bin'
 	];
+	if(process.env.V8_SNAPSHOT_ARGS) {
+		this.options.args = process.env.V8_SNAPSHOT_ARGS.split(/\s+/);
+	}
 	this.options.args.push(this.options.target || 'main.js');
 }
 module.exports = SnapshotPlugin;
 SnapshotPlugin.prototype.apply = function(compiler) {
 	var opts = this.options;
+	opts.blob = getBlobName(opts.args);
+
+	// Record the v8 blob file in the root appinfo if applicable
+	compiler.plugin('compilation', function(compilation) {
+		compilation.plugin('webos-meta-root-appinfo', function(meta) {
+			meta.v8SnapshotFile = opts.blob;
+			return meta;
+		});
+	});
 
 	compiler.plugin('after-emit', function(compilation, callback) {
 		if(isNodeOutputFS(compiler) && opts.exec) {
 			var ssCache = path.join(findCacheDir({
-					name: 'enact-dev',
-					create: true
-				}), 'snapshot-target.js');
+				name: 'enact-dev',
+				create: true
+			}), 'snapshot-target.js');
 
 			// Append anything optional to the js to be included in the snapshot
 			if(opts.prepend || opts.append) {
@@ -77,15 +74,13 @@ SnapshotPlugin.prototype.apply = function(compiler) {
 
 			if(child.status === 0) {
 				// Add snapshot to the compilation assets array for stats purposes
-				var blob = getBlobName(opts.args);
 				try {
-					var stat = fs.statSync(path.join(compiler.options.output.path, blob));
+					var stat = fs.statSync(path.join(compiler.options.output.path, opts.blob));
 					if(stat.size>0) {
-						compilation.assets[blob] = {
+						compilation.assets[opts.blob] = {
 							size: function() { return stat.size; },
 							emitted: true
 						};
-						err = updateAppInfo(compiler.options.output.path, blob);
 					} else {
 						// Temporary fix: mksnapshot may create a 0-byte blob on error
 						err = new Error(child.stdout + '\n' + child.stderr);