+++ global-cli/modifiers/util/snapshot-helper.js
@@ -1,12 +1,12 @@
 /*
  *  snapshot-helper.js
  *
- *  Since the initial ExecutionEnvironment is set during snapshot creation, we need to provide an API
- *  to update environment on window load.
- *
+ *  An exposed utility function to update the javascript environment to the active window to account for any
+ *  launch-time issues when using code created in a snapshot blob.
  */
 
 global.updateEnvironment = function() {
+	// Update fbjs to have the correct execution environment for the active window.
 	var ExecutionEnvironment = require('fbjs/lib/ExecutionEnvironment');
 	var canUseDOM = !!(typeof window !== 'undefined' && window.document && window.document.createElement);
 
@@ -15,4 +15,18 @@ global.updateEnvironment = function() {
 	ExecutionEnvironment.canUseEventListeners = canUseDOM && !!(window.addEventListener || window.attachEvent);
 	ExecutionEnvironment.canUseViewport = canUseDOM && !!window.screen;
 	ExecutionEnvironment.isInWorker = !canUseDOM; // For now, this is true - might change in the future.
+
+	// Mark the iLib localestorage cache as needing re-validation.
+	var ilib = require('@enact/i18n/ilib/lib/ilib');
+	if (ilib._load) {
+		ilib._load._cacheValidated = false;
+	}
+
+	// Clear the active resBundle and string cache.
+	var resBundle = require('@enact/i18n/src/resBundle');
+	resBundle.clearResBundle();
+
+	// Update the iLib/Enact locale to the active window's locale.
+	var locale = require('@enact/i18n/locale');
+	locale.updateLocale();
 };