+++ global-cli/pack.js
@@ -53,7 +53,6 @@ function printFileSizes(stats, previousSizeMap) {
 	var assets = stats.toJson().assets
 		.filter(asset => /\.(js|css|bin)$/.test(asset.name))
 		.map(asset => {
-			var fileContents = fs.readFileSync('./dist/' + asset.name);
 			var size = fs.statSync('./dist/' + asset.name).size;
 			var previousSize = previousSizeMap[shortFilename(asset.name)];
 			var difference = getDifferenceLabel(size, previousSize);
@@ -75,16 +74,14 @@ function printFileSizes(stats, previousSizeMap) {
 			var rightPadding = ' '.repeat(longestSizeLabelLength - sizeLength);
 			sizeLabel += rightPadding;
 		}
-		console.log(
-			'	' + sizeLabel +
-			'	' + chalk.dim(asset.folder + path.sep) + chalk.cyan(asset.name)
-		);
+		console.log('	' + sizeLabel +	'	' + chalk.dim(asset.folder + path.sep)
+				+ chalk.cyan(asset.name));
 	});
 }
 
 // Create the build and optionally, print the deployment instructions.
-function build(config, previousSizeMap, guided) {
-	if(process.env.NODE_ENV === 'development') {
+function build(config, previousSizeMap) {
+	if (process.env.NODE_ENV === 'development') {
 		console.log('Creating a development build...');
 	} else {
 		console.log('Creating an optimized production build...');
@@ -101,29 +98,22 @@ function build(config, previousSizeMap, guided) {
 		printFileSizes(stats, previousSizeMap);
 		console.log();
 		console.log(chalk.green('Compiled successfully.'));
-		console.log();
-
-		if(guided) {
-			var openCommand = process.platform === 'win32' ? 'start' : 'open';
-			console.log('The ' + chalk.cyan('dist') + ' directory is ready to be deployed.');
-			console.log('You may also serve it locally with a static server:');
-			console.log();
-			console.log('	' + chalk.cyan('npm') +	' install -g pushstate-server');
-			console.log('	' + chalk.cyan('pushstate-server') + ' dist');
-			console.log('	' + chalk.cyan(openCommand) + ' http://localhost:9000');
-			console.log();
+		if (process.env.NODE_ENV === 'development') {
+			console.log(chalk.yellow('NOTICE: This build contains debugging functionality and may run'
+					+ ' slower than in production mode.'));
 		}
+		console.log();
 	});
 }
 
 // Create the build and watch for changes.
 function watch(config) {
-	if(process.env.NODE_ENV === 'development') {
+	if (process.env.NODE_ENV === 'development') {
 		console.log('Creating a development build and watching for changes...');
 	} else {
 		console.log('Creating an optimized production build and watching for changes...');
 	}
-	webpack(config).watch({}, (err, stats) => {
+	webpack(config).watch({}, (err) => {
 		if (err) {
 			console.error('Failed to create ' + process.env.NODE_ENV + ' build. Reason:');
 			console.error(err.message || err);
@@ -142,7 +132,7 @@ function displayHelp() {
 	console.log('    -w, --watch       Rebuild on file changes');
 	console.log('    -p, --production  Build in production mode');
 	console.log('    -i, --isomorphic  Use isomorphic code layout');
-	console.log('                      (Includes prerendering)');
+	console.log('                      (includes prerendering)');
 	console.log('    -v, --version     Display version information');
 	console.log('    -h, --help        Display help information');
 	console.log();
@@ -153,6 +143,12 @@ function displayHelp() {
 			--framework           Builds the @enact/*, react, and react-dom into an external framework
 			--externals           Specify a local directory path to the standalone external framework
 			--externals-inject    Remote public path to the external framework for use injecting into HTML
+			--locales             Specifies to prerender locales. Can be:
+			                          "used" - Prerender locales used within ./resources/
+			                          "tv" - Prerender locales supported on the TV platform
+			                          "signage" - Prerender locales supported on the signage platform
+			                          "ilib" - Prerender all locales that iLib supports
+			                          <commana-separated-values> - Prerender the specifically listed locales
 	*/
 	process.exit(0);
 }
@@ -160,17 +156,17 @@ function displayHelp() {
 module.exports = function(args) {
 	var opts = minimist(args, {
 		boolean: ['minify', 'framework', 's', 'stats', 'p', 'production', 'i', 'isomorphic', 'snapshot', 'w', 'watch', 'h', 'help'],
-		string: ['externals', 'externals-inject'],
+		string: ['externals', 'externals-inject', 'locales'],
 		default: {minify:true},
 		alias: {s:'stats', p:'production', i:'isomorphic', w:'watch', h:'help'}
 	});
-	opts.help && displayHelp();
+	if (opts.help) displayHelp();
 
 	process.env.NODE_ENV = 'development';
 	var config = devConfig;
 
 	// Do this as the first thing so that any code reading it knows the right env.
-	if(opts.production) {
+	if (opts.production) {
 		process.env.NODE_ENV = 'production';
 		config = prodConfig;
 	}
@@ -178,11 +174,11 @@ module.exports = function(args) {
 	modifiers.apply(config, opts);
 
 	// Warn and crash if required files are missing
-	if (!opts.framework && !checkRequiredFiles([config.entry.main[config.entry.main.length-1]])) {
+	if (!opts.framework && !checkRequiredFiles([config.entry.main[config.entry.main.length - 1]])) {
 		process.exit(1);
 	}
 
-	if(opts.watch) {
+	if (opts.watch) {
 		watch(config);
 	} else {
 		// Read the current file sizes in dist directory.
@@ -200,7 +196,7 @@ module.exports = function(args) {
 			// if you're in it, you don't end up in Trash
 			try {
 				rimrafSync('dist/*');
-			} catch(e) {
+			} catch (e) {
 				console.log(chalk.red('Error: ') + ' Unable to delete existing build files. '
 						+ 'Please close any programs currently accessing files within ./dist/.');
 				console.log();